The database provider is now eagerly initialized.
This shifts the database initialization cost to the predictable app startup time instead of being paid unpredictably at the first instance where the database is needed.

This is done because this database is a major feature of this app and is expected to be needed every time this app is started up.